+++ ChangeLog.md
@@ -1,5 +1,13 @@
 # Dis ChangeLog
 
+## v 0.5.0
+
+### Framework
+
+* dis.assertion_summary - view of each assertion state
+* wrap sql/complete/dis.sql with a begin/commit block
+* add CONTRIBUTING.md
+
 ## v 0.4.0
 
 ### Framework
@@ -130,7 +138,7 @@
 * Management
 ** bin/schema - extract code from database
 ** bin/gensql - package sql for updates
-* Summary Views: 
+* Summary Views:
   * dis.schema_summary - summarize last results by schema
   * dis.module_summary - summarize last results by module
   * dis.submodule_summary - summarize last results by submodule